+++ lib/async/internal/blocking_operation_wait_inline.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+# Released under the MIT License.
+# Copyright, 2025, by Samuel Williams.
+
+module Async
+	module Internal
+		module BlockingOperationWaitInline
+			def blocking_operation_wait(work)
+				Fiber.blocking{work.call}
+			end
+		end
+	end
+end
